requirements.txt
Pillow>=5.1.0,<6
requests>=2.18.4,<3
